Update Data Handling Flow
This flow handles the gathering of system data from the OPC UA servers and their insertion into a session based SQLite database. Each PLC has its own 'Stream'.

There are 3 Stream types:

'Normal' Streams (for PLCs 09, 10, 11, 12, 13):
- These streams take a value change on the xCarrierAvailable identifier as a trigger.
- This trigger passes to the Triple Trigger Subflow and a link-out to a read request in the corresponding OPC UA Server Subflow via the OPC UA Server Flow.
- The Triple Trigger Subflow triggers three messages; a JS timestamp, a Process Type (the PLC number), and the original trigger value with msg.topic attributes corresponding to the column names for these values in the SQLite database (Timestamp, ProcessType, xCarrierAvailable). NB: xCarrierAvailable is not used as a database column, it is kept for sorting logic later in the stream.
- The read request results in a message generated at an MQTT subscription node containing the OrderID value, this passes a Change node that sets the msg.topic to OrderID, corresponding to the database column name for this value.
- The resulting messages from the Triple Trigger Subflow and the read request are then routed to the SQL Insert/Update Logic Subflow, where they are joined into one message object of key:value pairs and routed based on the value of the xCarrierAvailable key-value. NB: The prior msg.topic is used as key for the msg.payload.
These new message objects then go through one of two possible operations:

1. If xCarrierAvailable = 1, then the TimeStamp, ProcessType, and OrderID values are injected into the SQL database as StartTime, ProcessType and OrderID values.
2. If xCarrierAvailable = 0, then the TimeStamp is updated into the SQL database as FinishTime, for the entry where ProcessType and OrderID match the existing values.

The End-of-Line Stream (for PLC 08):
- As above, but a function node is added to read and remember the OrderID value for the xCarrierAvailable = 0 message object, as this value is unassigned at some point during the function of this Station. The function reads the OrderID number, and either saves it to a flow. context variable if it is not 0, or replaces the OrderID with the previously stored variable if it is 0 and the previous OrderID value was not 0. 
NB: This flow is not yet successfully recording cycle times.

The Complex Station Streams (for PLCs 01 & 14):
- These streams mirror the 'normal' streams but are doubled up, due to the additional stopping places on each of these modules. 
NB: These flows are not yet successfully recording cycle times, and have additionally not been tested.

NB: The delay times were a method for solving an issue with wrongfully attributed OrderID values. The 'normal' flows have been successfully tested with these delays with up to 3 pallets in the system. Following some testing, it seems likely that the root cause is that the PLCs only update their OrderID value at the end of their process cycle. Additionally, the success rate of records is inversely related to the number of pallets in the system, indicating a possible traffic issue on the network.