Create CycleTimesDisplay
This flow handles the display of data gathered to the database via tables on the Node-RED dashboard.

Each table has the same display functions:
- An inject node initialises the display at startup, sending a the following message to the database, the result of which is displayed chronologically via a table template:
SELECT * FROM (table name) ORDER BY (StartTime for processes, TimeStamp for orders) DESC;
- Each table has a search function enabled by a Text Input node, a Switch node, and a series of Change nodes which construct an SQL Select command, with the text input as a conditional value (i.e. 'SELECT * WHERE OrderID = INPUT ORDER BY StartTime DESC;'). In case of a null value, the original Select command is sent (see point immediately prior) - i.e. automatic refresh if the search bar is blank. 
- The process table has an update button implemented, which sends the original Select command on click (see above). This is yet to be implemented in the orders table.

There is also a number of nodes below these, which are the result of attempts to generate an automatically exported .CSV-file on button click. Currently, the file is generated on the Raspberry Pi. The http nodes are part of an attempt to implement an automatic download to the client computer on click of the corresponding button.