Add files via upload
This file is in Danish. It is the bachelor report that I submitted based on my work on this demonstrator. It contains a more in-depth description of node configurations and motivations/reasoning for these. The appedices (Bilag) contain debug reports and UI outputs for several tests carried out on the system.